refactor(handy-core): move audio_toolkit::text into handy-core
