+++ crates/handy-core/src/lib.rs
@@ -4,6 +4,9 @@
 //! `gtk-*`, or any GUI/platform-specific library. Platform capabilities are
 //! injected through traits defined in the `handy-platform` crate.
 
+pub mod text;
+pub use text::{apply_custom_words, filter_transcription_output};
+
 #[cfg(test)]
 mod sanity_check {
     #[test]
