+++ src-tauri/Cargo.toml
@@ -26,6 +26,7 @@ serde_json = "1"
 serde = { version = "1", features = ["derive"] }
 
 [dependencies]
+handy-core = { path = "../crates/handy-core" }
 once_cell = "1"
 tauri = { version = "2.10.2", features = [
   "protocol-asset",
